requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
PyJWT==2.8.0
joblib==1.3.2
bcrypt==4.0.1
scikit-learn==1.2.2
numpy==1.24.3
